+++ src/utilities.ts
@@ -25,7 +25,9 @@ export function estimateRunTime(command: string) {
 function runOptions(command: string, folder: string) {
 	const env = { ...process.env };
 	const javaHome: string = vscode.workspace.getConfiguration('ionic').get('javaHome');
-	if (command.includes('sync')) {
+
+	// Cocoapods required lang set to en_US.UTF-8 (when capacitor sync or run ios is done)
+	if (command.includes('sync') || command.includes('cap run ios')) {
 		env.LANG = 'en_US.UTF-8';
 	}
 	if (javaHome) {